src/promedic/requirements.txt
Django==1.10.5
djangorestframework==3.4.0
django-cors-middleware==1.3.1
#MySQL-python==1.2.3
mysqlclient==1.4.2
djangorestframework-jwt==1.9.0
gunicorn
twilio
Pillow
django-fsm
authy